==> 3_balancing.ipynb <==
import pandas as pd

X_df = pd.read_csv('data/dataset/dfu_features_dataset.csv', index_col=0)
y_df = pd.read_csv('data/dataset/dfu_labels_dataset.csv', index_col=0)

from imblearn.over_sampling import SMOTE
from collections import Counter

X = X_df.to_numpy()
y = y_df.to_numpy().ravel()
print('Original dataset shape %s' % Counter(y))

sm = SMOTE(random_state=42)
X_res, y_res = sm.fit_resample(X, y)
print('Resampled dataset shape %s' % Counter(y_res))

# saving results in csv files, whith same index and same columns names
X_res_df = pd.DataFrame(X_res, columns=X_df.columns)
y_res_df = pd.DataFrame(y_res, columns=y_df.columns)

X_res_df.to_csv('data/dataset/oversampling/dfu_features_dataset.csv')
y_res_df.to_csv('data/dataset/oversampling/dfu_labels_dataset.csv')

X_selected_df = pd.read_csv('data/dataset/dfu_features_dataset_selected.csv', index_col=0)
X_res_df[X_selected_df.columns].to_csv('data/dataset/oversampling/dfu_features_dataset_selected.csv')
